# Conflicts:
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Context/BookStoreContext.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170711134917_BookStore.Designer.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170711134917_BookStore.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170725194021_initial.Designer.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170725194021_initial.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170728155307_bookstore1.Designer.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/20170728155307_bookstore1.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/BookStoreContextModelSnapshot.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Repo/IRepo.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.Web/LvivCompany.Bookstore.Web.csproj
#	LvivCompany.Bookstore/LvivCompany.Bookstore.Web/Startup.cs
#	LvivCompany.Bookstore/LvivCompany.Bookstore.Web/Views/AddBook/AddBook.cshtml
#	LvivCompany.Bookstore/LvivCompany.Bookstore.Web/appsettings.json

+++ LvivCompany.Bookstore/LvivCompany.Bookstore.DataAccess/Migrations/BookStoreContextModelSnapshot.cs
@@ -61,7 +61,7 @@ namespace LvivCompany.Bookstore.DataAccess.Migrations
 
                     b.Property<long>("SellerId");
 
-                    b.Property<short>("Year");
+                    b.Property<int>("Year");
 
                     b.HasKey("Id");
